# cmd/swaggerui/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "GastroGang API v1",
    "version": "1.0.0",
    "title": "GastroGang"
  },
  "basePath": "/api/v1",
...
  "schemes": [
    "https",
    "http"
  ],
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "paths": {
...
    "/recipes/{id}": {
      "get": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "Returns a user based on a single ID, if the user does not have access to the pet",
        "summary": "Fetch recipe by ID",
        "tags": [
          "recipe"
        ],
        "operationId": "findRecipeByID",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of recipe to fetch",
            "required": true,
            "type": "integer",
            "format": "uint"
          }
        ],
        "responses": {
          "200": {
            "description": "Recipe",
            "schema": {
              "$ref": "#/definitions/Recipe"
            }
          },
          "403": {
            "description": "Recipe belongs to someone else"
          }
        }
      },
      "delete": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "Delete recipe by ID.",
        "summary": "Delete recipe by ID",
        "tags": [
          "recipe"
        ],
        "operationId": "deleteRecipeByID",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of recipe to fetch",
            "required": true,
            "type": "integer",
            "format": "uint"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "403": {
            "description": "Recipe belongs to someone else"
          },
          "404": {
            "description": "Recipe doesnt exist"
          }
        }
      },
      "put": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "tags": [
          "recipe"
        ],
        "summary": "Update recipe",
        "description": "",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of recipe to update",
            "required": true,
            "type": "integer",
            "format": "uint"
          },
          {
            "in": "body",
            "name": "recipe",
            "description": "Recipe to save",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                },
                "steps": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                "ingredients": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                "details": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "operationId": "saveRecipe",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Invalid body format"
          },
          "403": {
            "description": "Recipe belongs to someone else"
          },
          "422": {
            "description": "Duplicate recipe name error, user already has a recipe with the same name."
          }
        }
      }
    },
...
  },
  "definitions": {
    "Recipe": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "uint"
        },
        "authorId": {
          "type": "integer",
          "format": "uint"
        },
        "name": {
          "type": "string",
          "format": "string"
        },
        "steps": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "details": {
          "type": "string",
          "format": "string"
        },
        "ingredients": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "tags": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "ispublic": {
          "type": "boolean",
          "default": false
        },
        "like": {
          "$ref": "#/definitions/Like"
        }
      }
    },
...
    "Like": {
      "type": "object",
      "properties": {
        "count": {
          "type": "integer"
        },
        "AuthorIDs": {
          "type": "array",
          "items": {
            "type": "integer"
          }
        }
      }
    },
...
  }
}